+++ riscof/Templates/setup/model/riscof_model.py
@@ -15,14 +15,19 @@ from riscof.pluginTemplate import pluginTemplate
 
 logger = logging.getLogger()
 
-class spike_simple(pluginTemplate):
-    __model__ = "Spike"
-    __version__ = "0.5.0"
+class dutname(pluginTemplate):
+    __model__ = "dutname"
+
+    #TODO: please update the below to indicate family, version, etc of your DUT.
+    __version__ = "XXX"
 
     def __init__(self, *args, **kwargs):
         sclass = super().__init__(*args, **kwargs)
 
         config = kwargs.get('config')
+
+        self.dut_exe = os.path.join(config['PATH'] if 'PATH' in config else "","dutname")
+        self.num_jobs = str(config['jobs'] if 'jobs' in config else 1)
         if config is None:
             print("Please enter input file paths in configuration.")
             raise SystemExit
@@ -33,29 +38,66 @@ class spike_simple(pluginTemplate):
 
         return sclass
 
-    def initialise(self, suite, work_dir, compliance_env):
+    def initialise(self, suite, work_dir, archtest_env):
+        if shutil.which(self.dut_exe) is None:
+            logger.error(self.dut_exe+' Not Found')
+            logger.error('Please install Executable for dutname to proceed further')
+            sys.exit(0)
         self.work_dir = work_dir
-        self.compile_cmd = ''
+
+        #TODO: The following assumes you are using the riscv-gcc toolchain. If
+        #      not please change appropriately
+        self.compile_cmd = 'riscv{1}-unknown-elf-gcc -march={0} \
+         -static -mcmodel=medany -fvisibility=hidden -nostdlib -nostartfiles\
+         -T '+self.pluginpath+'/env/link.ld\
+         -I '+self.pluginpath+'/env/\
+         -I ' + archtest_env
 
         # set all the necessary variables like compile command, elf2hex
         # commands, objdump cmds. etc whichever you feel necessary and required
         # for your plugin. 
 
     def build(self, isa_yaml, platform_yaml):
-        ispec = utils.load_yaml(isa_yaml)
-        self.isa = ispec["ISA"]
+        ispec = utils.load_yaml(isa_yaml)['hart0']
+        self.xlen = ('64' if 64 in ispec['supported_xlen'] else '32')
+        self.isa = 'rv' + self.xlen
+        #TODO: The following assumes you are using the riscv-gcc toolchain. If
+        #      not please change appropriately
+        self.compile_cmd = self.compile_cmd+' -mabi='+('lp64 ' if 64 in ispec['supported_xlen'] else 'ilp32 ')
+        if "I" in ispec["ISA"]:
+            self.isa += 'i'
+        if "M" in ispec["ISA"]:
+            self.isa += 'm'
+        if "C" in ispec["ISA"]:
+            self.isa += 'c'
 
-        # based on the validated isa and platform configure your simulator ro
+        # based on the validated isa and platform configure your simulator or
         # build your RTL here
 
-    def runTests(self, testList):
+    def runTests(self, testList, cgf_file=None):
+        make = utils.makeUtil(makefilePath=os.path.join(self.work_dir, "Makefile." + self.name[:-1]))
+        make.makeCommand = 'make -j' + self.num_jobs
         for file in testList:
             testentry = testList[file]
-            test = os.path.join(constants.root, str(file))
+            test = testentry['test_path']
             test_dir = testentry['work_dir']
 
-            # compile each test
-            # execute each test on simulator/DUT
-            # fix signature format if required
+            elf = 'dut.elf'
             
+            execute = "@cd "+testentry['work_dir']+";"
 
+            cmd = self.compile_cmd.format(testentry['isa'].lower(), self.xlen) + ' ' + test + ' -o ' + elf
+
+            #TODO: we are using -D to enable compile time macros. If your
+            #      toolchain is not riscv-gcc you may want to change the below code
+            compile_cmd = cmd + ' -D' + " -D".join(testentry['macros'])
+            execute+=compile_cmd+";"
+
+            sig_file = os.path.join(test_dir, self.name[:-1] + ".signature")
+
+            #TODO: You will need to add any other arguments to your DUT
+            #      executable if any in the quotes below
+            execute += self.dut_exe + '' 
+
+            make.add_target(execute)
+        make.execute_all(self.work_dir)